fix TECH-576
When a backup path is provided, if multiple validation attempts are made,
the transport system always issues a mkdir. This causes dropbox to throw
"409 path/conflict/folder/". To fix this, we capture the error by localizing
STDERR and should "error_summary" match "^path/conflict/folder" on mkdir, we
ignore it.